feat(*): split services mock file
ref: ISSUE-13805
