+++ packages/manager/apps/hpc-vmware-managed-vcd/src/test-utils/render-test.tsx
@@ -12,7 +12,7 @@ import { render, waitFor, screen } from '@testing-library/react';
 import {
   getServicesMocks,
   GetServicesMocksParams,
-} from '@ovh-ux/manager-react-components/src/hooks/services/mocks/services.mock';
+} from '@ovh-ux/manager-react-components/src/hooks/services/mocks';
 import { toMswHandlers } from '../../../../../../playwright-helpers';
 import { getAuthenticationMocks } from '../../../../../../playwright-helpers/mocks/auth';
 import {
