+++ packages/manager/apps/veeam-backup/tests/helpers/setup-test.tsx
@@ -11,7 +11,7 @@ import { render, waitFor, screen } from '@testing-library/react';
 import {
   getServicesMocks,
   GetServicesMocksParams,
-} from '@ovh-ux/manager-react-components/src/hooks/services/mocks/services.mock';
+} from '@ovh-ux/manager-react-components/src/hooks/services/mocks';
 import { toMswHandlers } from '../../../../../../playwright-helpers';
 import { getAuthenticationMocks } from '../../../../../../playwright-helpers/mocks/auth';
 import {
